// src/lib/type-defs.graphqls.d.ts
export declare type Maybe<T> = T | null;
export declare type Exact<T extends {
    [key: string]: unknown;
}> = {
    [K in keyof T]: T[K];
};
export declare type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
    [SubKey in K]?: Maybe<T[SubKey]>;
};
export declare type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
    [SubKey in K]: Maybe<T[SubKey]>;
};
export declare type RequireFields<T, K extends keyof T> = {
    [X in Exclude<keyof T, K>]?: T[X];
} & {
    [P in K]-?: NonNullable<T[P]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export declare type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
};
export declare type Auth = {
    __typename?: 'Auth';
    isAuthenticated: Scalars['Boolean'];
    userFullName: Scalars['String'];
};
export declare type Query = {
    __typename?: 'Query';
    viewer: OldUser;
    auth: Auth;
};
export declare type Mutation = {
    __typename?: 'Mutation';
    updateName: OldUser;
    signIn: Auth;
};
export declare type MutationUpdateNameArgs = {
    name: Scalars['String'];
};
export declare type MutationSignInArgs = {
    name?: Maybe<Scalars['String']>;
};
export declare type OldUser = {
    __typename?: 'OldUser';
    id: Scalars['ID'];
    name: Scalars['String'];
    status: Scalars['String'];
};
export declare type ResolverTypeWrapper<T> = Promise<T> | T;
export declare type LegacyStitchingResolver<TResult, TParent, TContext, TArgs> = {
    fragment: string;
    resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export declare type NewStitchingResolver<TResult, TParent, TContext, TArgs> = {
    selectionSet: string;
    resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export declare type StitchingResolver<TResult, TParent, TContext, TArgs> = LegacyStitchingResolver<TResult, TParent, TContext, TArgs> | NewStitchingResolver<TResult, TParent, TContext, TArgs>;
export declare type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | StitchingResolver<TResult, TParent, TContext, TArgs>;
export declare type ResolverFn<TResult, TParent, TContext, TArgs> = (parent: TParent, args: TArgs, context: TContext, info: GraphQLResolveInfo) => Promise<TResult> | TResult;
// ...
export declare type ResolversTypes = {
    Auth: ResolverTypeWrapper<Auth>;
    Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
    String: ResolverTypeWrapper<Scalars['String']>;
    Query: ResolverTypeWrapper<{}>;
    Mutation: ResolverTypeWrapper<{}>;
    OldUser: ResolverTypeWrapper<OldUser>;
    ID: ResolverTypeWrapper<Scalars['ID']>;
};
/** Mapping between all available schema types and the resolvers parents */
export declare type ResolversParentTypes = {
    Auth: Auth;
    Boolean: Scalars['Boolean'];
    String: Scalars['String'];
    Query: {};
    Mutation: {};
    OldUser: OldUser;
    ID: Scalars['ID'];
};
export declare type AuthResolvers<ContextType = any, ParentType extends ResolversParentTypes['Auth'] = ResolversParentTypes['Auth']> = {
    isAuthenticated?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
    userFullName?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};
export declare type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
    viewer?: Resolver<ResolversTypes['OldUser'], ParentType, ContextType>;
    auth?: Resolver<ResolversTypes['Auth'], ParentType, ContextType>;
};
export declare type MutationResolvers<ContextType = any, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
    updateName?: Resolver<ResolversTypes['OldUser'], ParentType, ContextType, RequireFields<MutationUpdateNameArgs, 'name'>>;
    signIn?: Resolver<ResolversTypes['Auth'], ParentType, ContextType, RequireFields<MutationSignInArgs, never>>;
};